+++ pypoman/polyhedron.py
@@ -67,7 +67,7 @@ def compute_chebyshev_center(A: np.ndarray, b: np.ndarray) -> np.ndarray:
     A_cheby = np.hstack([A, a_cheby.reshape((A.shape[0], 1))])
     z = solve_lp(cost, A_cheby, b)
     if z[-1] < -1e-1:  # last coordinate is distance to boundaries
-        raise Exception("Polytope is empty (margin violation %.2f)" % z[-1])
+        raise ValueError("Polytope is empty (margin violation %.2f)" % z[-1])
     return z[:-1]