+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
+- Raise ``ValueError`` exception when polyhedron is empty
 - Remove default QP solver when projecting a point to a polytope
 
 ## [1.0.0] - 2023/05/18